requirements.txt
textual
telethon
python-dotenv